IGNITE-17824 / move JumpToExecutorByConsistentId to network-api
